Swap imgui library

+++ ContentEditor.App/Imgui/Common/BasicImguiHandlers.cs
@@ -1,7 +1,6 @@
 using System.Numerics;
 using ContentEditor.App.Windowing;
 using ContentPatcher;
-using ImGuiNET;
 using ReeLib.Aimp;
 using ReeLib.Common;
 using ReeLib.via;
@@ -13,7 +12,7 @@ public class NumericFieldHandler<T>(ImGuiDataType type) : IObjectUIHandler where
     public unsafe void OnIMGUI(UIContext context)
     {
         var num = (T)context.Get<object>();
-        if (ImGui.DragScalar(context.label, type, (IntPtr)(&num), type is ImGuiDataType.Float or ImGuiDataType.Double ? 0.01f : 0.05f)) {
+        if (ImGui.DragScalar(context.label, type, &num, type is ImGuiDataType.Float or ImGuiDataType.Double ? 0.01f : 0.05f)) {
             UndoRedo.RecordSet(context, num);
         }
     }
